Start Owncast channel from template
Following the guidance on how to use Roku's template at
https://github.com/rokudev/rdp-to-scenegraph-channel-template,
the manifest was updated with Owncast's Roku feed.

Additionally, since Owncast only uses "live feeds", the logic in
RootHandler was changed to create an "all streams" section as well as
sections for "Music", "Tech", "Chatting", and "Video Games".  More
may be added as appropriate, and likely this logic can be leveraged
to create "recently live" and other sections.