Add various graphical elements and theming
This commit adds a splash screen, a main menu icon, a backgroung,
and a logo for the "overhang", which sits at the top of the screen.

* The origin of the overhang is logo-glare-outlined.png in the Owncast
design kit.
* The origin of the background is owncast-background, also from the
design kit.
* Original source for the splash and icon came from Gabe Kangas, who
shared it from an icon pack for the iOS/tvOS Owncast client.

Right now, the theming is kinda all over the place.  The main page
has the "classic" owncast theming, while the focus icon and splash
are from something more novel.  The color schemes don't really line
up.  Someone with better design skills could help with this.