Show all tags for all feeds
This commit copies a string joining function to RootHandler so that all the
tags are joined into a comma-separated string.  Then, a change in DetailsView.xml
and DetailsView.brs allows flowing the tags over 4 lines of text (to keep a
vertical alignment with the max length of a channel description).

This lets us show all the tags for the feeds rather than just the first.

Note that it's not clear we actually want to show these, but it looks better
than when it was showing just the first one.