RootHandler.brs: fix bug causing only first tag to be considered
For some reason, the loop checking the tags for each feed ws only looking
at the first tag.  This was a "for each" style loop and should operate on
an roArray just fine, but for some inexplicable reason, it was not.  Moving
this over to a traditional counter-style loop solved the problem.

Maybe someone smarter at BrightScript will look at this and know why, but
it's beyond the scope of my ability for an eary bring-up of the app.