components MainScene: load DetailsView only for selected item
In the default app template, selecting an item from the MainScene loads
its entire row into the DetailsView, which treats the entire row as a
playlist.  This would mean when a stream went offline, it would auto-
play the next stream in the row.

Rather than refactor out the entire "row as playlist" metaphor, this
commit hacks a simple solution-- load only the selected item into the
DetailsView and specify "not a content list", which locks down the UI
to working with only the selected stream and not the entire row.

Resolves issue #9.