Add "favorites" feature and badging for live streams
This commit introduces two features: favorites and badging for live
feeds.  These are being added together in one commit because the former
really strongly depends on the latter.  That is, once a user has
favorites, wanting to know if a favorite is live is essential.

The UI flow is very simple and works with what's already present in the
app.  When selecting a feed and going to its details page, there is the
option to add it as a favorite or remove it from favorites if it was
previously added.  Additionaly, the "Favorites" row appears only if the
user has selected some favorites.

This is a feature that's going to have some jank to it.  It only stores
the ID from the directory for a feed and relies on its directory entry
to gather details.  If a stream disappears from the directory, it'll
disappear from the user's favorites, too.  This could mean a slight
leakage of storage space and a little UI jank if someone favorites a
feed that disappears for months.  But, this can be cleaned up with
future work.

Additionally, because the "Favorites" row will contain a mix of live and
not-live streams, this commit adds a "LIVE" badge to any stream that has
the "now-live" tag, making it easier to browse favorites and topical
rows for the currently live streams.  Probably, in the future, we should
move all the live feeds to the front of their respective rows.